Model refactoring: remove unused methods and classes
