+++ src/main/java/it/unibo/jurassiko/model/territory/api/BoardArea.java
@@ -17,26 +17,6 @@ public interface BoardArea<T extends BoardArea<T>> {
     /**
      * @return the name of the neighbouring areas
      */
-    Set<String> getNeighbourNames();
-
-    /**
-     * Sets the neighbours on the specific board area.
-     * 
-     * @param neighbours
-     */
-    void setNeighbours(Set<T> neighbours);
-
-    /**
-     * @return the instance of the neighbours
-     */
-    Set<T> getNeighbours();
-
-    /**
-     * Checks if the given area is a neighbour.
-     * 
-     * @param name the name of the area to check
-     * @return true if the area is a neighbour, false otherwise
-     */
-    boolean isNeighbour(String name);
+    Set<String> getNeighbours();
 
 }
