+++ src/main/java/it/unibo/jurassiko/model/territory/impl/OceanFactoryImpl.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 import it.unibo.jurassiko.model.territory.api.Ocean;
 import it.unibo.jurassiko.model.territory.api.OceanFactory;
-import it.unibo.jurassiko.reader.impl.OceanReader;
+import it.unibo.jurassiko.reader.impl.BoardDataReader;
 
 /**
  * Implementation of the factory for the game oceans.
@@ -18,7 +18,7 @@ public class OceanFactoryImpl implements OceanFactory {
      */
     @Override
     public Set<Ocean> createOceans() {
-        final var oceanReader = new OceanReader();
+        final var oceanReader = new BoardDataReader<>(Ocean.class);
         final Set<Ocean> oceans = oceanReader.readFileData(PATH);
         return Set.copyOf(oceans);
     }
