+++ src/main/java/it/unibo/jurassiko/model/territory/impl/OceanImpl.java
@@ -1,15 +1,11 @@
 package it.unibo.jurassiko.model.territory.impl;
 
-import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import it.unibo.jurassiko.model.territory.api.BoardArea;
 import it.unibo.jurassiko.model.territory.api.Ocean;
-import it.unibo.jurassiko.model.territory.api.Territory;
 
 /**
  * Implementation of the game Ocean.
@@ -18,38 +14,18 @@ import it.unibo.jurassiko.model.territory.api.Territory;
 public final class OceanImpl extends AbstractBoardArea<Ocean> implements Ocean {
 
     @JsonProperty("territories")
-    private Set<String> adjTerritoryNames;
-
-    @JsonIgnore
-    private Set<Territory> adjTerritories;
+    private Set<String> adjTerritories;
 
     // NOPMD using the class for Jackson deserializer for the factory
     private OceanImpl() {
-        this.adjTerritories = new HashSet<>();
+
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public Set<String> getAdjTerritoryNames() {
-        return Set.copyOf(adjTerritoryNames);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setAdjTerritories(final Set<Territory> adjTerritories) {
-        Objects.requireNonNull(adjTerritories);
-        this.adjTerritories = new HashSet<>(adjTerritories);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public Set<Territory> getAdjTerritories() {
+    public Set<String> getAdjTerritories() {
         return Set.copyOf(adjTerritories);
     }
 
@@ -60,7 +36,6 @@ public final class OceanImpl extends AbstractBoardArea<Ocean> implements Ocean {
     public boolean isAdjTerritory(final String territoryName) {
         Objects.requireNonNull(adjTerritories);
         return adjTerritories.stream()
-                .map(BoardArea::getName)
                 .anyMatch(territoryName::equals);
     }
