+++ src/main/java/it/unibo/jurassiko/model/territory/impl/TerritoryFactoryImpl.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 import it.unibo.jurassiko.model.territory.api.Territory;
 import it.unibo.jurassiko.model.territory.api.TerritoryFactory;
-import it.unibo.jurassiko.reader.impl.TerritoryReader;
+import it.unibo.jurassiko.reader.impl.BoardDataReader;
 
 /**
  * Implementation of the factory for the game territories.
@@ -18,7 +18,7 @@ public class TerritoryFactoryImpl implements TerritoryFactory {
      */
     @Override
     public Set<Territory> createTerritories() {
-        final var territoryReader = new TerritoryReader();
+        final var territoryReader = new BoardDataReader<>(Territory.class);
         final Set<Territory> territories = territoryReader.readFileData(PATH);
         return Set.copyOf(territories);
     }
