--- src/main/java/it/unibo/jurassiko/reader/impl/TerritoryReader.java
@@ -1,27 +1,0 @@
-package it.unibo.jurassiko.reader.impl;
-
-import java.util.Set;
-
-import it.unibo.jurassiko.model.territory.api.Territory;
-
-/**
- * Implementation of JSONFileReader for territories.
- */
-public class TerritoryReader extends AbstractBoardDataReader<Territory> {
-
-    /**
-     * Creates a TerritoryReader.
-     */
-    public TerritoryReader() {
-        super(Territory.class);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected void buildAttributes(final Set<Territory> data) {
-        defineNeighbours(data);
-    }
-
-}
