+++ src/test/java/it/unibo/jurassiko/TestTerritory.java
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -35,6 +34,7 @@ class TestTerritory {
         assertNotNull(territories);
         assertFalse(territories.isEmpty());
 
+        // Tests the amount of territories per continent
         assertEquals(NUM_TOTAL_TERRITORIES, territories.size());
         assertEquals(NUM_TOTAL_TERRITORIES, territories.stream().map(Territory::getName).distinct().count());
         assertEquals(NUM_CONTINENTS,
@@ -53,16 +53,12 @@ class TestTerritory {
     void testTerritoryAttributes() {
         final String name = "Nord Africa";
         final String continent = "Gondwana Occidentale";
-        final var neighbourNames = Set.of("Arabia", "Amazzonia", "Congo", "Sud America");
-        final var neighbours = territories.stream()
-                .filter(t -> neighbourNames.contains(t.getName()))
-                .collect(Collectors.toSet());
+        final var neighbours = Set.of("Arabia", "Amazzonia", "Congo", "Sud America");
 
         assertTrue(territories.stream().anyMatch(t -> t.getName().equals(name)));
 
         final Territory sampleTerritory = territories.stream().filter(t -> t.getName().equals(name)).findAny().get();
         assertEquals(continent, sampleTerritory.getContinent());
-        assertEquals(neighbourNames, sampleTerritory.getNeighbourNames());
         assertEquals(neighbours, sampleTerritory.getNeighbours());
     }
